adk-agent/requirements.txt
# Google ADK and AI
google-adk>=0.3.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.38.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Firebase
firebase-admin>=6.4.0

# HTTP client
httpx>=0.26.0
aiohttp>=3.9.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment
python-dotenv>=1.0.0

# Utilities
orjson>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
